+++ demos/python/sdk_wireless_camera_control/open_gopro/demos/gui/livestream.py
@@ -73,15 +73,8 @@ def parse_arguments() -> argparse.Namespace:
     parser.add_argument("--min_bit", type=int, help="Minimum bitrate.", default=1000)
     parser.add_argument("--max_bit", type=int, help="Maximum bitrate.", default=1000)
     parser.add_argument("--start_bit", type=int, help="Starting bitrate.", default=1000)
-    parser.add_argument(
-        "--resolution",
-        help="Resolution.",
-        choices=list(proto.EnumWindowSize.values()),
-        default=proto.EnumWindowSize.WINDOW_SIZE_720,
-    )
-    parser.add_argument(
-        "--fov", help="Field of View.", choices=list(proto.EnumLens.values()), default=proto.EnumLens.LENS_LINEAR
-    )
+    parser.add_argument("--resolution", help="Resolution.", choices=list(proto.EnumWindowSize.values()), default=None)
+    parser.add_argument("--fov", help="Field of View.", choices=list(proto.EnumLens.values()), default=None)
     return add_cli_args_and_parse(parser, wifi=False)